demo bao cao

+++ src/main/java/Models/NhiemVuBaoTriDAO.java
@@ -70,30 +70,32 @@ public class NhiemVuBaoTriDAO {
         
         StringBuilder query = new StringBuilder("insert into NhiemVubaoTri(iD, iDKeHoachBaoTri, tenNhiemVu, chiTietNhiemVu, danhGia, chiPhi) values ");
         ArrayList<NhiemVuBaoTri> khbts = (ArrayList<NhiemVuBaoTri>) nhiemVuBaoTris;
-        khbts.forEach(nhiemVuBaoTri -> {
-            query.append("(?, ?, ?, ?, ?, ?),");
-        });
-        query.deleteCharAt(query.length() - 1);
-        
-        pttm = conn.prepareStatement(query.toString());
-        
-        // Khoảng cách các dấu hỏi
-        int distance = 6;
-        
-        // Thêm các giá trị cho dãy truy vẫn
-        for(int i=0; i<khbts.size(); i++) {
-            NhiemVuBaoTri nhiemVuBaoTriChild = khbts.get(i);
-            pttm.setString((i * distance) + 1, nhiemVuBaoTriChild.getiD());
-            pttm.setString((i * distance) + 2, nhiemVuBaoTriChild.getiDKeHoachBaoTri());
-            pttm.setString((i * distance) + 3, nhiemVuBaoTriChild.getTenNhiemVu());
-            pttm.setString((i * distance) + 4, nhiemVuBaoTriChild.getChiTietNhiemVu());
-            pttm.setString((i * distance) + 5, nhiemVuBaoTriChild.getDanhGia());
-            pttm.setInt((i * distance) + 6, nhiemVuBaoTriChild.getChiPhi());
-        }
-        
-        if(pttm.executeUpdate() > 0) {
-            System.out.println("Insert thanh cong");
-            return 1;
+        if(khbts != null && !khbts.isEmpty()) {
+            khbts.forEach(nhiemVuBaoTri -> {
+                query.append("(?, ?, ?, ?, ?, ?),");
+            });
+            query.deleteCharAt(query.length() - 1);
+
+            pttm = conn.prepareStatement(query.toString());
+
+            // Khoảng cách các dấu hỏi
+            int distance = 6;
+
+            // Thêm các giá trị cho dãy truy vẫn
+            for(int i=0; i<khbts.size(); i++) {
+                NhiemVuBaoTri nhiemVuBaoTriChild = khbts.get(i);
+                pttm.setString((i * distance) + 1, nhiemVuBaoTriChild.getiD());
+                pttm.setString((i * distance) + 2, nhiemVuBaoTriChild.getiDKeHoachBaoTri());
+                pttm.setString((i * distance) + 3, nhiemVuBaoTriChild.getTenNhiemVu());
+                pttm.setString((i * distance) + 4, nhiemVuBaoTriChild.getChiTietNhiemVu());
+                pttm.setString((i * distance) + 5, nhiemVuBaoTriChild.getDanhGia());
+                pttm.setInt((i * distance) + 6, nhiemVuBaoTriChild.getChiPhi());
+            }
+
+            if(pttm.executeUpdate() > 0) {
+                System.out.println("Insert thanh cong");
+                return 1;
+            }
         }
         return -1;
     }
